Python-Utils/requirements.txt
python-dotenv==1.1.0
sib-api-v3-sdk==7.6.0
requests==2.31.0 